refactor apiserver start to avoid multiple overwrites and side channels
